refactor: move functions inside script
